test(runner): cover deployment inventory boundaries
